Move URL normalization logic to shared util
